refactor. 컴포넌트 분리, 밖 클릭시 주소 창 닫히도록 설정
- 슬라이도 피드백 반영하여 외부 클릭시 주소창이 닫히게 바꿨습니다
- DB 업데이트 단위로 분리하여 채널 활성화와 삭제 영역을 ActiveChannel / DeleteChannel 로 분리했습니다
